Implemented the queue of event results' processing in client; implemented movement and camera animations.

+++ src/org/tendiwa/client/CharacterActor.java
@@ -6,7 +6,6 @@ import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import tendiwa.core.Character;
-import tendiwa.core.Tendiwa;
 
 public class CharacterActor extends Actor {
 private static final TextureAtlas atlasCharacters = new TextureAtlas(Gdx.files.internal("pack/characters.atlas"), true);
@@ -20,14 +19,14 @@ public CharacterActor(Character character) {
 	String type = "player";
 	texture = atlasCharacters.findRegion(type);
 	if (texture == null) {
-		throw new RuntimeException("No image for character type "+ type);
+		throw new RuntimeException("No image for character type " + type);
 	}
 
 }
 
 @Override
 public void draw(SpriteBatch batch, float parentAlpha) {
-	batch.draw(texture, getX()*GameScreen.TILE_SIZE, getY()* GameScreen.TILE_SIZE);
+	batch.draw(texture, (int) (getX() * GameScreen.TILE_SIZE), (int) (getY() * GameScreen.TILE_SIZE));
 }
 
 }
